+++ scripts/assets/README.md
